Real time messaging fixes; legend
Fixed several issues with the real time data handling. Tested using
virtual hospital which was awesome. Added a basic legend and
the styling changes for notification of a comment. Also fixed
notifications to not build up when a message id is specified.

AIRFLOW-15 #ready #comment Made a basic legend for the status bars only
ARIFLOW-20 #ready #comment Comments now stand on their own in the notification area