Fix additional event tab headings getting added on new events
Because the event nav list was in a handlebars tempate, when a new event
occurred and the area was re-rendered, an additional nav header was getting
added. This is resolved by moving the nav outside of the template.

Additionally, when new events occur and the template is re-rendered, it switches
to the comments pane. Moving the nav outside the template, it does not change to
reflect the active pane. This is resolved by adding the active class to the
comment nav when a new event is added.