Add background color when hovering a card over a lane
To show which resource it is going to get dropped on.

Adds additional check to prevent unnecessary network requests by making sure a
lane change or time change occurred. There's an (arbitrary) 5 pixel +/- change
required to indicate a time change. These checks are in place mainly to prevent
an adjustment getting made when a card is picked up and dropped back in the same place.