Refactor items in resource selector to be more React-y
Made each item in the resource select as it's own component, in order to
pass the resourceName into the onClick function, while avoiding the
previous method of getting the name from a data field on the html
element. Creating a new component was necessary to avoid having an arrow
function in the onClick, which causes a new bind on every rerender.