Removed need for static absolute pos after rotating Overview headers
The Overview lane headers were being rotated backwards 90 degrees via
the right side, which then required positioning the element to the left
and up to compensate. We are now instead using writing mode to make the
text vertical, but then we still have to rotate 180 degrees (which
doesn't require positioning the element afterwords).

This should all work in IE9, but put it in a different commit just in
case.