Refactor lib/[utility | data].js files and add lib/index.js
Moves functions around between these two files and the new selectors.js to try
to be more clear about how and why functions are located where they are.

An index is added for lib to make imports cleaner in components, e.g. allowing a
single import for files in lib instead of multiple imports.