Fix client requests when recv messages for non-app orders
Messages that were coming in on v1.1.2 were causing requests to /exam_info for
all orders. The check for whether to fetch data was resulting in cases where the
order was not in the app to always equal undefined. Additionally, the check for
which resource the order belonged to was returning undefined for new orders (not
attached to an exam). This resulted in the original behavior of updates not
appearing in the app.