Fix potential issue with incorrect order resource selected
Previously, when getting the resource for an order (which only happens
when creating the unique indentifier used to group orders into cards),
the first resource that matches either the adjusted resource id or rad
exam resource id is chosen, even if the first one that matches is the
rad exam resource id. Now, the adjusted resource id always takes
precedent.